edit#3 merge and clean up unnecessary files.
