+++ app/src/main/java/com/example/todoapp/ui/main/main/MainViewModel.kt
@@ -3,10 +3,11 @@ package com.example.todoapp.ui.main.main
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import com.example.todoapp.dao.UserDao
 import com.example.todoapp.ui.main.util.Destination
 import com.example.todoapp.ui.main.util.Event
 
-class MainViewModel : ViewModel() {
+class MainViewModel(val userDao: UserDao) : ViewModel() {
 
     private val _destination = MutableLiveData<Event<Destination>>()
     val destination: LiveData<Event<Destination>>
