+++ app/src/main/java/com/example/todoapp/ui/main/register/RegisterFragment.kt
@@ -1,4 +1,4 @@
-package com.example.todoapp.ui.main
+package com.example.todoapp.ui.main.register
 
 import android.os.Bundle
 import android.view.View
