+++ app/src/main/java/com/example/todoapp/ui/main/register/RegisterViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.todoapp.ui.main
+package com.example.todoapp.ui.main.register
 
 import androidx.lifecycle.ViewModel
